Support assigning a json string to a mixed property
Collect all to_json related functions in one compilation unit. Then
it will only be included in the final binary if used.
