+++ src/realm/obj.cpp
@@ -1104,151 +1104,6 @@ void Obj::add_index(Path& path, Index index) const
     }
 }
 
-void Obj::to_json(std::ostream& out, size_t link_depth, const std::map<std::string, std::string>& renames,
-                  std::vector<ObjLink>& followed, JSONOutputMode output_mode) const
-{
-    followed.push_back(get_link());
-    size_t new_depth = link_depth == not_found ? not_found : link_depth - 1;
-    StringData name = "_key";
-    bool prefixComma = false;
-    if (renames.count(name))
-        name = renames.at(name);
-    out << "{";
-    if (output_mode == output_mode_json) {
-        prefixComma = true;
-        out << "\"" << name << "\":" << this->m_key.value;
-    }
-
-    auto col_keys = m_table->get_column_keys();
-    for (auto ck : col_keys) {
-        name = m_table->get_column_name(ck);
-        auto type = ck.get_type();
-        if (type == col_type_LinkList)
-            type = col_type_Link;
-        if (renames.count(name))
-            name = renames.at(name);
-
-        if (prefixComma)
-            out << ",";
-        out << "\"" << name << "\":";
-        prefixComma = true;
-
-        TableRef target_table;
-        ColKey pk_col_key;
-        if (type == col_type_Link) {
-            target_table = get_target_table(ck);
-            pk_col_key = target_table->get_primary_key_column();
-        }
-
-        auto print_link = [&](const Mixed& val) {
-            REALM_ASSERT(val.is_type(type_Link, type_TypedLink));
-            TableRef tt = target_table;
-            auto obj_key = val.get<ObjKey>();
-            std::string table_info;
-            std::string table_info_close;
-            if (!tt) {
-                // It must be a typed link
-                tt = m_table->get_parent_group()->get_table(val.get_link().get_table_key());
-                pk_col_key = tt->get_primary_key_column();
-                if (output_mode == output_mode_xjson_plus) {
-                    table_info = std::string("{ \"$link\": ");
-                    table_info_close = " }";
-                }
-
-                table_info += std::string("{ \"table\": \"") + std::string(tt->get_name()) + "\", \"key\": ";
-                table_info_close += " }";
-            }
-            if (pk_col_key && output_mode != output_mode_json) {
-                out << table_info;
-                tt->get_primary_key(obj_key).to_json(out, output_mode_xjson);
-                out << table_info_close;
-            }
-            else {
-                ObjLink link(tt->get_key(), obj_key);
-                if (obj_key.is_unresolved()) {
-                    out << "null";
-                    return;
-                }
-                if (!tt->is_embedded()) {
-                    if (link_depth == 0) {
-                        out << table_info << obj_key.value << table_info_close;
-                        return;
-                    }
-                    if ((link_depth == realm::npos &&
-                         std::find(followed.begin(), followed.end(), link) != followed.end())) {
-                        // We have detected a cycle in links
-                        out << "{ \"table\": \"" << tt->get_name() << "\", \"key\": " << obj_key.value << " }";
-                        return;
-                    }
-                }
-
-                tt->get_object(obj_key).to_json(out, new_depth, renames, followed, output_mode);
-            }
-        };
-
-        if (ck.is_collection()) {
-            if (m_table->get_nesting_levels(ck)) {
-                auto collection_list = get_collection_list(ck);
-                collection_list->to_json(out, link_depth, output_mode, print_link);
-            }
-            else {
-                auto collection = get_collection_ptr(ck);
-                collection->to_json(out, link_depth, output_mode, print_link);
-            }
-        }
-        else {
-            auto val = get_any(ck);
-            if (!val.is_null()) {
-                if (type == col_type_Link) {
-                    std::string close_string;
-                    bool is_embedded = target_table->is_embedded();
-                    bool link_depth_reached = !is_embedded && (link_depth == 0);
-
-                    if (output_mode == output_mode_xjson_plus) {
-                        out << "{ " << (is_embedded ? "\"$embedded" : "\"$link") << "\": ";
-                        close_string += "}";
-                    }
-                    if ((link_depth_reached && output_mode == output_mode_json) ||
-                        output_mode == output_mode_xjson_plus) {
-                        out << "{ \"table\": \"" << target_table->get_name() << "\", "
-                            << (is_embedded ? "\"value" : "\"key") << "\": ";
-                        close_string += "}";
-                    }
-
-                    print_link(val);
-                    out << close_string;
-                }
-                else if (val.is_type(type_TypedLink)) {
-                    print_link(val);
-                }
-                else if (val.is_type(type_Dictionary)) {
-                    DummyParent parent(m_table, val.get_ref());
-                    Dictionary dict(parent, 0);
-                    dict.to_json(out, link_depth, output_mode, print_link);
-                }
-                else if (val.is_type(type_Set)) {
-                    DummyParent parent(this->get_table(), val.get_ref());
-                    Set<Mixed> set(parent, 0);
-                    set.to_json(out, link_depth, output_mode, print_link);
-                }
-                else if (val.is_type(type_List)) {
-                    DummyParent parent(m_table, val.get_ref());
-                    Lst<Mixed> list(parent, 0);
-                    list.to_json(out, link_depth, output_mode, print_link);
-                }
-                else {
-                    val.to_json(out, output_mode);
-                }
-            }
-            else {
-                out << "null";
-            }
-        }
-    }
-    out << "}";
-    followed.pop_back();
-}
-
 std::string Obj::to_string() const
 {
     std::ostringstream ostr;
